Move remote_listener to test_api
