+++ pkgs/test/pubspec.yaml
@@ -1,5 +1,5 @@
 name: test
-version: 1.5.0
+version: 1.5.1
 author: Dart Team <misc@dartlang.org>
 description: A full featured library for writing and running Dart tests.
 homepage: https://github.com/dart-lang/test/pkgs/test
@@ -30,8 +30,8 @@ dependencies:
   web_socket_channel: '^1.0.0'
   yaml: '^2.0.0'
   # Use an exact version until the test_api and test_core package are stable.
-  test_api: '0.2.0'
-  test_core: '0.1.0'
+  test_api: '0.2.1'
+  test_core: '0.2.0'
 dev_dependencies:
   fake_async: '>=0.1.2 <2.0.0'
   pedantic: '1.1.0'
