+++ pkgs/test_api/lib/src/suite_channel_manager.dart
