+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Remote `remote_listener.dart` and `suite_channel_manager.dart` from runner
+  and depend on them from `test_api`. 
+
+
 ## 0.1.0
 
 * Initial release of `test_core`. Provides the basic API for writing and running
